+++ quadra/utils/utils.py
@@ -31,8 +31,6 @@ from quadra.utils.mlflow import infer_signature_torch_model
 
 IMAGE_EXTENSIONS: List[str] = [".png", ".jpg", ".jpeg", ".bmp", ".tiff", ".tif", ".pbm", ".pgm", ".ppm", ".pxm", ".pnm"]
 
-IMAGE_EXTENSIONS: List[str] = [".png", ".jpg", ".jpeg", ".bmp", ".tiff", ".tif", ".pbm", ".pgm", ".ppm", ".pxm", ".pnm"]
-
 
 def get_logger(name=__name__) -> logging.Logger:
     """Initializes multi-GPU-friendly python logger."""
@@ -181,11 +179,19 @@ def log_hyperparameters(
         )
         == 0
     ):
-        hparams["git/commit"] = subprocess.check_output(["git", "rev-parse", "HEAD"]).decode("ascii").strip()
-        hparams["git/branch"] = (
-            subprocess.check_output(["git", "rev-parse", "--abbrev-ref", "HEAD"]).decode("ascii").strip()
-        )
-        hparams["git/remote"] = subprocess.check_output(["git", "remote", "get-url", "origin"]).decode("ascii").strip()
+        try:
+            hparams["git/commit"] = subprocess.check_output(["git", "rev-parse", "HEAD"]).decode("ascii").strip()
+            hparams["git/branch"] = (
+                subprocess.check_output(["git", "rev-parse", "--abbrev-ref", "HEAD"]).decode("ascii").strip()
+            )
+            hparams["git/remote"] = (
+                subprocess.check_output(["git", "remote", "get-url", "origin"]).decode("ascii").strip()
+            )
+        except subprocess.CalledProcessError:
+            log.warning(
+                "Could not get git commit, branch or remote information, the repository might not have any commits yet "
+                "or it might be initialized wrongly."
+            )
     else:
         log.warning("Could not find git repository, skipping git commit and branch info")
 
@@ -270,7 +276,9 @@ def finish(
             if model_json is not None:
                 for model_path in deployed_models:
                     if model_path.endswith(".pt"):
-                        model, _ = quadra_export.import_deployment_model(model_path, device="cpu")
+                        model = quadra_export.import_deployment_model(
+                            model_path, device="cpu", inference_config=config.inference
+                        ).model
 
                         input_size = model_json["input_size"]