+++ tests/tasks/test_anomaly.py
@@ -2,14 +2,24 @@
 import os
 import shutil
 from pathlib import Path
+from typing import List
 
 import pytest
 
-from quadra.utils.tests.fixtures import base_anomaly_dataset
-from quadra.utils.tests.helpers import execute_quadra_experiment
+from quadra.utils.export import get_export_extension
+from quadra.utils.tests.fixtures import base_anomaly_dataset, imagenette_dataset
+from quadra.utils.tests.helpers import check_deployment_model, execute_quadra_experiment, setup_trainer_for_lightning
+
+try:
+    import onnx  # noqa
+    import onnxruntime  # noqa
+    import onnxsim  # noqa
+
+    ONNX_AVAILABLE = True
+except ImportError:
+    ONNX_AVAILABLE = False
 
 BASE_EXPERIMENT_OVERRIDES = [
-    "trainer=lightning_cpu",
     "trainer.devices=1",
     "datamodule.num_workers=1",
     "datamodule.train_batch_size=1",
@@ -20,19 +30,7 @@ BASE_EXPERIMENT_OVERRIDES = [
     "~logger.mlflow",
 ]
 
-
-def _check_deployment_model(invert: bool = False):
-    """Check that the runtime model is present and valid.
-
-    Args:
-        invert: If true check that the runtime model is not present.
-    """
-    if invert:
-        assert not os.path.exists("deployment_model/model.pt")
-        assert not os.path.exists("deployment_model/model.json")
-    else:
-        assert os.path.exists("deployment_model/model.pt")
-        assert os.path.exists("deployment_model/model.json")
+BASE_EXPORT_TYPES = ["torchscript"] if not ONNX_AVAILABLE else ["torchscript", "onnx"]
 
 
 def _check_report(invert: bool = False):
@@ -51,7 +49,10 @@ def _check_report(invert: bool = False):
         assert os.path.exists("avg_score_by_label.csv")
 
 
-def _run_inference_experiment(data_path: str, train_path: str, test_path: str):
+def _run_inference_experiment(data_path: str, train_path: str, test_path: str, export_type: str):
+    """Run an inference experiment for the given export type."""
+    extension = get_export_extension(export_type)
+
     test_overrides = [
         "task.device=cpu",
         "experiment=base/anomaly/inference",
@@ -59,14 +60,29 @@ def _run_inference_experiment(data_path: str, train_path: str, test_path: str):
         "datamodule.num_workers=1",
         "datamodule.test_batch_size=32",
         "logger=csv",
-        f"task.model_path={os.path.join(train_path, 'deployment_model', 'model.pt')}",
+        f"task.model_path={os.path.join(train_path, 'deployment_model', f'model.{extension}')}",
     ]
 
     execute_quadra_experiment(overrides=test_overrides, experiment_path=test_path)
 
 
+def run_inference_experiments(data_path: str, train_path: str, test_path: str, export_types: List[str]):
+    """Run inference experiments for the given export types."""
+    for export_type in export_types:
+        cwd = os.getcwd()
+        check_deployment_model(export_type=export_type)
+
+        _run_inference_experiment(
+            data_path=data_path, train_path=train_path, test_path=test_path, export_type=export_type
+        )
+
+        # Change back to the original working directory
+        os.chdir(cwd)
+
+
 @pytest.mark.parametrize("task", ["classification", "segmentation"])
-def test_padim_training(tmp_path: Path, base_anomaly_dataset: base_anomaly_dataset, task: str):
+def test_padim(tmp_path: Path, base_anomaly_dataset: base_anomaly_dataset, task: str):
+    """Test the training and evaluation of the PADIM model."""
     data_path, _ = base_anomaly_dataset
 
     train_path = tmp_path / "train"
@@ -77,23 +93,28 @@ def test_padim_training(tmp_path: Path, base_anomaly_dataset: base_anomaly_datas
         f"datamodule.data_path={data_path}",
         "model.model.backbone=resnet18",
         f"model.dataset.task={task}",
+        f"export.types=[{','.join(BASE_EXPORT_TYPES)}]",
     ]
+    trainer_overrides = setup_trainer_for_lightning()
     overrides += BASE_EXPERIMENT_OVERRIDES
+    overrides += trainer_overrides
 
     execute_quadra_experiment(overrides=overrides, experiment_path=train_path)
 
     assert os.path.exists("checkpoints/final_model.ckpt")
 
-    _check_deployment_model()
     _check_report()
 
-    _run_inference_experiment(data_path, train_path, test_path)
+    run_inference_experiments(
+        data_path=data_path, train_path=train_path, test_path=test_path, export_types=BASE_EXPORT_TYPES
+    )
 
     shutil.rmtree(tmp_path)
 
 
 @pytest.mark.parametrize("task", ["classification", "segmentation"])
-def test_patchcore_training(tmp_path: Path, base_anomaly_dataset: base_anomaly_dataset, task: str):
+def test_patchcore(tmp_path: Path, base_anomaly_dataset: base_anomaly_dataset, task: str):
+    """Test the training and evaluation of the PatchCore model."""
     data_path, _ = base_anomaly_dataset
 
     train_path = tmp_path / "train"
@@ -104,24 +125,70 @@ def test_patchcore_training(tmp_path: Path, base_anomaly_dataset: base_anomaly_d
         f"datamodule.data_path={data_path}",
         "model.model.backbone=resnet18",
         f"model.dataset.task={task}",
+        f"export.types=[{','.join(BASE_EXPORT_TYPES)}]",
     ]
+    trainer_overrides = setup_trainer_for_lightning()
     overrides += BASE_EXPERIMENT_OVERRIDES
+    overrides += trainer_overrides
 
     execute_quadra_experiment(overrides=overrides, experiment_path=train_path)
 
     assert os.path.exists("checkpoints/final_model.ckpt")
 
-    _check_deployment_model()
     _check_report()
 
-    _run_inference_experiment(data_path, train_path, test_path)
+    run_inference_experiments(
+        data_path=data_path, train_path=train_path, test_path=test_path, export_types=BASE_EXPORT_TYPES
+    )
+
+    shutil.rmtree(tmp_path)
+
+
+@pytest.mark.parametrize("task", ["classification", "segmentation"])
+def test_efficientad(
+    tmp_path: Path, base_anomaly_dataset: base_anomaly_dataset, imagenette_dataset: imagenette_dataset, task: str
+):
+    """Test the training and evaluation of the EfficientAD model."""
+    data_path, _ = base_anomaly_dataset
+    imagenette_path = imagenette_dataset
+
+    train_path = tmp_path / "train"
+    test_path = tmp_path / "test"
+
+    overrides = [
+        "experiment=base/anomaly/efficient_ad",
+        f"datamodule.data_path={data_path}",
+        "transforms.input_height=256",
+        "transforms.input_width=256",
+        "model.model.train_batch_size=1",
+        "datamodule.test_batch_size=1",
+        "model.model.image_size=[256, 256]",
+        "trainer.check_val_every_n_epoch= ${trainer.max_epochs}",
+        f"model.model.imagenette_dir= {imagenette_path}",
+        f"model.dataset.task={task}",
+        f"export.types=[{','.join(BASE_EXPORT_TYPES)}]",
+    ]
+    trainer_overrides = setup_trainer_for_lightning()
+    overrides += BASE_EXPERIMENT_OVERRIDES
+    overrides += trainer_overrides
+
+    execute_quadra_experiment(overrides=overrides, experiment_path=train_path)
+
+    assert os.path.exists("checkpoints/final_model.ckpt")
+
+    _check_report()
+
+    run_inference_experiments(
+        data_path=data_path, train_path=train_path, test_path=test_path, export_types=BASE_EXPORT_TYPES
+    )
 
     shutil.rmtree(tmp_path)
 
 
 @pytest.mark.slow
 @pytest.mark.parametrize("task", ["classification", "segmentation"])
-def test_cflow_training(tmp_path: Path, base_anomaly_dataset: base_anomaly_dataset, task: str):
+def test_cflow(tmp_path: Path, base_anomaly_dataset: base_anomaly_dataset, task: str):
+    """Test the training and evaluation of the cflow model."""
     data_path, _ = base_anomaly_dataset
 
     overrides = [
@@ -129,24 +196,26 @@ def test_cflow_training(tmp_path: Path, base_anomaly_dataset: base_anomaly_datas
         f"datamodule.data_path={data_path}",
         "model.model.backbone=resnet18",
         f"model.dataset.task={task}",
-        "task.export_config=null",
+        "export.types=[]",
     ]
+    trainer_overrides = setup_trainer_for_lightning()
     overrides += BASE_EXPERIMENT_OVERRIDES
+    overrides += trainer_overrides
 
     execute_quadra_experiment(overrides=overrides, experiment_path=tmp_path)
 
     assert os.path.exists("checkpoints/final_model.ckpt")
 
-    _check_deployment_model(invert=True)  # cflow does not support runtime model
     _check_report()
 
-    # cflow does not support inference with jitted model
+    # cflow does not support exporting to torchscript and onnx at the moment
     shutil.rmtree(tmp_path)
 
 
 @pytest.mark.slow
 @pytest.mark.parametrize("task", ["classification", "segmentation"])
-def test_csflow_training(tmp_path: Path, base_anomaly_dataset: base_anomaly_dataset, task: str):
+def test_csflow(tmp_path: Path, base_anomaly_dataset: base_anomaly_dataset, task: str):
+    """Test the training and evaluation of the csflow model."""
     data_path, _ = base_anomaly_dataset
     train_path = tmp_path / "train"
     test_path = tmp_path / "test"
@@ -155,51 +224,63 @@ def test_csflow_training(tmp_path: Path, base_anomaly_dataset: base_anomaly_data
         "experiment=base/anomaly/csflow",
         f"datamodule.data_path={data_path}",
         f"model.dataset.task={task}",
+        f"export.types=[{','.join(BASE_EXPORT_TYPES)}]",
     ]
+    trainer_overrides = setup_trainer_for_lightning()
     overrides += BASE_EXPERIMENT_OVERRIDES
+    overrides += trainer_overrides
 
     execute_quadra_experiment(overrides=overrides, experiment_path=train_path)
 
     assert os.path.exists("checkpoints/final_model.ckpt")
 
-    _check_deployment_model()
     _check_report()
 
-    _run_inference_experiment(data_path, train_path, test_path)
+    run_inference_experiments(
+        data_path=data_path, train_path=train_path, test_path=test_path, export_types=BASE_EXPORT_TYPES
+    )
 
     shutil.rmtree(tmp_path)
 
 
 @pytest.mark.slow
 @pytest.mark.parametrize("task", ["classification", "segmentation"])
-def test_fastflow_training(tmp_path: Path, base_anomaly_dataset: base_anomaly_dataset, task: str):
+def test_fastflow(tmp_path: Path, base_anomaly_dataset: base_anomaly_dataset, task: str):
+    """Test the training and evaluation of the fastflow model."""
     data_path, _ = base_anomaly_dataset
     train_path = tmp_path / "train"
     test_path = tmp_path / "test"
+
+    export_types = ["torchscript"]  # fastflow does not support exporting to onnx at the moment
 
     overrides = [
         "experiment=base/anomaly/fastflow",
         f"datamodule.data_path={data_path}",
         "model.model.backbone=resnet18",
         f"model.dataset.task={task}",
+        f"export.types=[{','.join(export_types)}]",
     ]
+    trainer_overrides = setup_trainer_for_lightning()
     overrides += BASE_EXPERIMENT_OVERRIDES
+    overrides += trainer_overrides
 
     execute_quadra_experiment(overrides=overrides, experiment_path=train_path)
 
     assert os.path.exists("checkpoints/final_model.ckpt")
 
-    _check_deployment_model()
     _check_report()
 
-    _run_inference_experiment(data_path, train_path, test_path)
+    run_inference_experiments(
+        data_path=data_path, train_path=train_path, test_path=test_path, export_types=export_types
+    )
 
     shutil.rmtree(tmp_path)
 
 
 @pytest.mark.slow
 @pytest.mark.parametrize("task", ["classification", "segmentation"])
-def test_draem_training(tmp_path: Path, base_anomaly_dataset: base_anomaly_dataset, task: str):
+def test_draem(tmp_path: Path, base_anomaly_dataset: base_anomaly_dataset, task: str):
+    """Test the training and evaluation of the draem model."""
     data_path, _ = base_anomaly_dataset
     train_path = tmp_path / "train"
     test_path = tmp_path / "test"
@@ -208,16 +289,20 @@ def test_draem_training(tmp_path: Path, base_anomaly_dataset: base_anomaly_datas
         "experiment=base/anomaly/draem",
         f"datamodule.data_path={data_path}",
         f"model.dataset.task={task}",
+        f"export.types=[{','.join(BASE_EXPORT_TYPES)}]",
     ]
+    trainer_overrides = setup_trainer_for_lightning()
     overrides += BASE_EXPERIMENT_OVERRIDES
+    overrides += trainer_overrides
 
     execute_quadra_experiment(overrides=overrides, experiment_path=train_path)
 
     assert os.path.exists("checkpoints/final_model.ckpt")
 
-    _check_deployment_model()
     _check_report()
 
-    _run_inference_experiment(data_path, train_path, test_path)
+    run_inference_experiments(
+        data_path=data_path, train_path=train_path, test_path=test_path, export_types=BASE_EXPORT_TYPES
+    )
 
     shutil.rmtree(tmp_path)